Použítí MediaPlayer k hudbě na pozadí, navržení práce s databází(AppDatabase) a ukládání skóre do ní. Vypisování z databáze pomocí listu. AsyncTask při insertování.

+++ app/src/main/java/cer0387/projekt_sudoku/Vysledek.java
@@ -2,9 +2,9 @@ package cer0387.projekt_sudoku;
 
 public class Vysledek {
     public final String nick;
-    public final long skore;
+    public final int skore;
 
-    Vysledek(String nick, long skore) {
+    Vysledek(String nick, int skore) {
         this.nick = nick;
         this.skore = skore;
     }
